lectures/03/powergrid-sampler/requirements.txt
# HDFS
kafka-python-ng

# Webserver
Flask==3
Gunicorn==22.0.0
